fatih/requirements.txt
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
lxml==5.4.0
numpy==2.0.2
onvif_zeep==0.2.12
opencv-python-headless==4.11.0.86
platformdirs==4.3.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
urllib3==2.5.0
zeep==4.3.1
